+++ js/fetch.js
@@ -1,4 +1,5 @@
-import {GET_URL, SEND_URL} from './constants.js';
+const GET_URL = 'https://25.javascript.pages.academy/kekstagram/data';  // Адрес сервера для получения изображений
+const SEND_URL = 'https://25.javascript.pages.academy/kekstagram';      // Адрес сервера для отправки изображений
 
 
 /**
@@ -33,10 +34,6 @@ const sendData = (onSuccess, onFail, body) => {
   fetch(SEND_URL,
     {
       method: 'POST',
-      headers: {
-        'Content-Type': 'multipart/form-data'
-      },
-      mode: 'no-cors',
       body,
     },
   )
